Replaced NFT model to single file
