+++ FakeNFT/Models/Domain/NFT.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct NFT {
+    let name: String
+    let image: String
+    let rating: Int
+    let price: Float
+    let id: String
+}
